EnemyBuild
Enemy is built and ready for animation. It broke into two pieces, and im not sure why. But it is done and is grouped together. Just be sure to drag both pieces in at the same time and it will be all one piece. Also finished the mountains in the foreground.